Move installed modules code to utils (#2429)
Even though we're now using the `_get_installed_modules` function in many different places, it still lives in `sentry_sdk.integrations.modules`. With this change we move `_get_installed_modules` (and related helpers) to `utils.py` and introduce a new `package_version` helper function (also in `utils.py`) that finds out and parses the version of a package in one go.
