+++ sentry_sdk/integrations/flask.py
@@ -5,13 +5,12 @@ from sentry_sdk.hub import Hub, _should_send_default_pii
 from sentry_sdk.integrations import DidNotEnable, Integration
 from sentry_sdk.integrations._wsgi_common import RequestExtractor
 from sentry_sdk.integrations.wsgi import SentryWsgiMiddleware
-from sentry_sdk.integrations.modules import _get_installed_modules
 from sentry_sdk.scope import Scope
 from sentry_sdk.tracing import SOURCE_FOR_STYLE
 from sentry_sdk.utils import (
     capture_internal_exceptions,
     event_from_exception,
-    parse_version,
+    package_version,
 )
 
 if TYPE_CHECKING:
@@ -64,13 +63,10 @@ class FlaskIntegration(Integration):
     @staticmethod
     def setup_once():
         # type: () -> None
-
-        installed_packages = _get_installed_modules()
-        flask_version = installed_packages["flask"]
-        version = parse_version(flask_version)
+        version = package_version("flask")
 
         if version is None:
-            raise DidNotEnable("Unparsable Flask version: {}".format(flask_version))
+            raise DidNotEnable("Unparsable Flask version.")
 
         if version < (0, 10):
             raise DidNotEnable("Flask 0.10 or newer is required.")
