+++ sentry_sdk/integrations/opentelemetry/integration.py
@@ -9,8 +9,7 @@ from importlib import import_module
 from sentry_sdk.integrations import DidNotEnable, Integration
 from sentry_sdk.integrations.opentelemetry.span_processor import SentrySpanProcessor
 from sentry_sdk.integrations.opentelemetry.propagator import SentryPropagator
-from sentry_sdk.integrations.modules import _get_installed_modules
-from sentry_sdk.utils import logger
+from sentry_sdk.utils import logger, _get_installed_modules
 from sentry_sdk._types import TYPE_CHECKING
 
 try:
